removed saving cookies

+++ index.py
@@ -1,15 +1,18 @@
+"""main file for flask app"""
+import os
 from flask import Flask, request, jsonify, send_file
 from flask_cors import CORS
 import lib
-import os
+
 
 app = Flask(__name__)
 
-# Enable CORS
+# Allow requests from all origins
 CORS(app, resources={r"/*": {"origins": "*"}})
 
 @app.route('/', methods=['POST'])
 def download_video_mp4():
+    """route for home page"""
     data = request.json  
     url = data.get('url')
 
@@ -22,13 +25,13 @@ def download_video_mp4():
         # Call getVideoType in lib.py, which will handle the video download
         file_path = lib.getVideoType(url)
 
-        if file_path:
+        if file_path and os.path.exists(file_path):
             # Serve the file to the client
             response = send_file(file_path, as_attachment=True)
 
             # After serving the file, delete it from the server
             try:
-                os.remove(file_path)
+                os.remove(file_path)  # File deletion after serving
                 print(f"File {file_path} deleted from server")
             except Exception as e:
                 print(f"Error deleting file: {e}")
