+++ mp4.py
@@ -1,13 +1,14 @@
+"""mp4 function to download media file"""
+import os
 import yt_dlp
 import lib
-import os
 
 def download_video_mp4(url):
     """Download video using cookies from the Chrome session."""
     try:
         save_path = './'
         output_file = 'output_file.mp4'
-        new_output_path = os.path.join(save_path, output_file)  # File path without checking if it exists
+        new_output_path = lib.checkFileExists(save_path, output_file)
 
         # Get cookies from Chrome
         cookies_file = lib.get_cookies_from_chrome()
@@ -33,11 +34,8 @@ def download_video_mp4(url):
                 print(f"Error downloading video: {e}")
                 return None
 
-        # Cleanup: Delete the temporary cookies file
-        if os.path.exists(cookies_file):
+    finally:
+        # Cleanup: Delete the temporary cookies file if it exists
+        if cookies_file and os.path.exists(cookies_file):
             os.remove(cookies_file)
             print(f"Temporary cookies file {cookies_file} deleted")
-
-    except Exception as e:
-        print(f"An unexpected error occurred: {e}")
-        return None
